backend/requirements.txt
# Multi-Agent Negotiation Framework Dependencies with ADK Support

# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.35.0
pydantic==2.11.7
pydantic-settings==2.10.1

# Google ADK and AI Dependencies
google-adk==1.6.1
google-genai==1.25.0
google-cloud-aiplatform==1.103.0
google-auth==2.40.3
google-api-python-client==2.176.0

# Database and Memory
redis[hiredis]==5.2.1
chromadb==0.5.20

# HTTP and WebSocket
httpx==0.28.1
websockets==15.0.1
python-multipart==0.0.20

# Async and Utilities
asyncio-mqtt==0.16.2
tenacity==8.5.0
python-dateutil==2.9.0.post0

# Logging and Monitoring
loguru==0.7.3
opentelemetry-api==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-exporter-gcp-trace==1.9.0

# Configuration and Environment
python-dotenv==1.1.1
pyyaml==6.0.2

# Development and Testing (optional)
pytest==8.3.4
pytest-asyncio==0.25.0
httpx==0.28.1  # for testing
pytest-mock==3.14.0

# Security
cryptography==45.0.5
authlib==1.6.0

# Data Processing
numpy==2.3.1
pandas==2.2.3

# Optional: Additional AI/ML Libraries
# openai==1.58.1  # If you want to keep OpenAI as fallback
# anthropic==0.40.0  # If you want to support Claude

# Optional: Advanced Consensus Algorithms
# scipy==1.14.1  # For advanced mathematical consensus methods

# Optional: Visualization (for agent interaction graphs)
# plotly==5.24.1
# networkx==3.4.2

# Optional: Advanced Caching
# diskcache==5.6.3

# Optional: Advanced Serialization
# msgpack==1.1.0

# Optional: Database Migrations (if needed)
# alembic==1.14.0
# sqlalchemy==2.0.41 